fileproxy/pyproject.toml
[project]
name = "fileproxy"
version = "0.1.0"
description = ""
authors = [
    {name = "jamin",email = "jaminbecker@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=6.0.1,<7.0.0)",
    "cryptography (>=46.0.3,<47.0.0)",
    "psycopg[binary] (>=3.3.2,<4.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "fileproxy", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
